+++ project-base/src/Shopsys/ShopBundle/Model/Payment/PaymentData.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Shopsys\ShopBundle\Model\Payment;
+
+use Shopsys\FrameworkBundle\Model\Payment\PaymentData as BasePaymentData;
+
+class PaymentData extends BasePaymentData
+{
+    public function __construct()
+    {
+        parent::__construct();
+    }
+}