+++ project-base/tests/ShopBundle/Database/Model/Order/OrderTransportAndPaymentTest.php
@@ -2,13 +2,13 @@
 
 namespace Tests\ShopBundle\Database\Model\Order;
 
-use Shopsys\FrameworkBundle\Model\Payment\Payment;
-use Shopsys\FrameworkBundle\Model\Payment\PaymentDataFactory;
+use Shopsys\FrameworkBundle\Model\Payment\PaymentDataFactoryInterface;
 use Shopsys\FrameworkBundle\Model\Payment\PaymentFacade;
 use Shopsys\FrameworkBundle\Model\Pricing\Vat\Vat;
 use Shopsys\FrameworkBundle\Model\Pricing\Vat\VatData;
 use Shopsys\FrameworkBundle\Model\Transport\TransportDataFactoryInterface;
 use Shopsys\FrameworkBundle\Model\Transport\TransportFacade;
+use Shopsys\ShopBundle\Model\Payment\Payment;
 use Shopsys\ShopBundle\Model\Transport\Transport;
 use Tests\ShopBundle\Test\DatabaseTestCase;
 
@@ -426,11 +426,11 @@ class OrderTransportAndPaymentTest extends DatabaseTestCase
     }
 
     /**
-     * @return \Shopsys\FrameworkBundle\Model\Payment\PaymentDataFactory
+     * @return \Shopsys\ShopBundle\Model\Payment\PaymentDataFactory
      */
     public function getPaymentDataFactory()
     {
-        return $this->getContainer()->get(PaymentDataFactory::class);
+        return $this->getContainer()->get(PaymentDataFactoryInterface::class);
     }
 
     /**