+++ project-base/tests/ShopBundle/Database/Model/Payment/PaymentTest.php
@@ -2,12 +2,12 @@
 
 namespace Tests\ShopBundle\Database\Model\Payment;
 
-use Shopsys\FrameworkBundle\Model\Payment\Payment;
-use Shopsys\FrameworkBundle\Model\Payment\PaymentDataFactory;
+use Shopsys\FrameworkBundle\Model\Payment\PaymentDataFactoryInterface;
 use Shopsys\FrameworkBundle\Model\Pricing\Vat\Vat;
 use Shopsys\FrameworkBundle\Model\Pricing\Vat\VatData;
 use Shopsys\FrameworkBundle\Model\Transport\TransportDataFactoryInterface;
 use Shopsys\FrameworkBundle\Model\Transport\TransportFacade;
+use Shopsys\ShopBundle\Model\Payment\Payment;
 use Shopsys\ShopBundle\Model\Transport\Transport;
 use Tests\ShopBundle\Test\DatabaseTestCase;
 
@@ -15,8 +15,8 @@ class PaymentTest extends DatabaseTestCase
 {
     public function testRemoveTransportFromPaymentAfterDelete()
     {
-        $paymentDataFactory = $this->getContainer()->get(PaymentDataFactory::class);
-        /** @var \Shopsys\FrameworkBundle\Model\Payment\PaymentDataFactory $paymentDataFactory */
+        $paymentDataFactory = $this->getContainer()->get(PaymentDataFactoryInterface::class);
+        /** @var \Shopsys\ShopBundle\Model\Payment\PaymentDataFactory $paymentDataFactory */
         $transportDataFactory = $this->getContainer()->get(TransportDataFactoryInterface::class);
         /** @var \Shopsys\ShopBundle\Model\Transport\TransportDataFactory $transportDataFactory */
         $em = $this->getEntityManager();