+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - Transport is extended
     - Product is extended
     - Brand is extended
+    - Payment is extended
 
 ### [shopsys/shopsys]
 #### Added